requirements.txt
pycryptodome>=3.9.9
ansicolors>=1.1.8
pyopenssl>=20.0.1
requests>=2.22.0